refactor: Remove old WebSocketManager indexes
- Update some features about this file moving (websocket.py to server.py)
