+++ AloneChat/core/network/__init__.py
@@ -1,4 +1,3 @@
-from .websocket import WebSocketManager
 from .protocol import Message, MessageType
 
-__all__ = ['WebSocketManager', 'Message', 'MessageType']
+__all__ = ['Message', 'MessageType']
