+++ __main__.py
@@ -1,11 +1,10 @@
 import argparse
-import asyncio
 from AloneChat.start import client, server
 
 def main():
     parser = argparse.ArgumentParser(prog='AloneChat', description='AloneChat starter')
     subparsers = parser.add_subparsers(dest='command', required=True, help='Available commands')
-
+    # Define subparsers for server and client
     server_parser = subparsers.add_parser('server', help='Startup SERVER')
     server_parser.add_argument('--port', type=int, default=8765, help='SERVER port (default: 8080)')
     # server_parser.add_argument('--debug', action='store_true', help='Enable debug mode')
